added activate button and improving examples
