+++ examples/demo-active.html
@@ -1,7 +1,7 @@
 <html>
   <head>
     <title>An active web page</title>
-
+    <link rel="stylesheet" type="text/css" href="index.css" />
     <script type="text/x-thebe-config">
       {
         bootstrap: true,
@@ -14,8 +14,10 @@
         },
       }
     </script>
-    <script type="text/javascript" src="https://unpkg.com/thebe@latest"></script>
-
+    <script
+      type="text/javascript"
+      src="https://unpkg.com/thebe@latest"
+    ></script>
   </head>
   <body>
     <h1>My web page with code samples</h1>
@@ -34,6 +36,5 @@ x = np.linspace(0,10)
 plt.plot(x, np.sin(x))
 plt.plot(x, np.cos(x))
     </pre>
-
   </body>
 </html>
