+++ examples/demo-launch-button.html
@@ -1,0 +1,81 @@
+<html>
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title>Thebe Custom Launch Button</title>
+    <link rel="stylesheet" type="text/css" href="index.css" />
+
+    <!-- Thebe configuration !-->
+    <script type="text/x-thebe-config">
+      {
+        binderOptions: {
+          repo: "binder-examples/requirements"
+        },
+        requestKernel: true,
+        mountStatusWidget: true,
+      }
+    </script>
+
+    <!-- Demo function to load the thebe library and immediately bootstrap !-->
+    <script>
+      // Activate button function hook
+      function thebe_launch_latest() {
+        $("#launch-button").prop("disabled", true);
+        // Load the Thebe library
+        $.getScript("https://unpkg.com/thebe@latest")
+          .done(function (script, textStatus) {
+            console.log(thebelab);
+            thebelab.mountStatusWidget();
+            thebelab.bootstrap();
+            $("#launch-button").remove();
+          })
+          .fail(function (jqxhr, settings, exception) {
+            $("div.log").text("Could not fetch Thebe library.");
+            $("#launch-button").prop("disabled", false);
+          });
+      }
+    </script>
+
+    <!-- Dependency: jquery !-->
+    <script
+      src="https://code.jquery.com/jquery-3.2.1.min.js"
+      type="text/javascript"
+    ></script>
+    <style>
+      body: {
+        padding: 20px;
+      }
+    </style>
+  </head>
+  <body>
+    <h1>Example Launch Button</h1>
+    <p>
+      This page contains and example of how to add a button to your webpage
+      which will automatically fetch the latest thebe version, load it and
+      bootstrap the library. This is standalone code, please view the page
+      source to see the example.
+    </p>
+    <button
+      id="launch-button"
+      type="button"
+      class="my-launch-button"
+      onclick="thebe_launch_latest()"
+    >
+      Launch
+    </button>
+    <div class="thebe-status"></div>
+    <p>Once Thebe is active this cell will turn into the live cell:</p>
+    <pre data-executable>
+%matplotlib inline
+import numpy as np
+import matplotlib.pyplot as plt
+x = np.linspace(0,10)
+plt.plot(x, np.sin(x))
+plt.plot(x, np.cos(x))
+    </pre>
+    <p>
+      Note: this page also contains the built in status widget - see
+      <a href="./demo-status-widget.html">demo-status-widget.html</a> for more
+      info.
+    </p>
+  </body>
+</html>
