+++ examples/demo-static.html
@@ -1,6 +1,7 @@
 <html>
   <head>
     <title>A static web page</title>
+    <link rel="stylesheet" type="text/css" href="index.css" />
   </head>
   <body>
     <h1>My web page with code samples</h1>
@@ -19,6 +20,5 @@
       plt.plot(x, np.sin(x))
       plt.plot(x, np.cos(x))
     </pre>
-
   </body>
 </html>
