+++ examples/demo-status-widget.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <title>Demo Status Widget</title>
-    <link rel="stylesheet" href="main.css" />
+    <link rel="stylesheet" type="text/css" href="index.css" />
     <!-- Configure and load Thebe !-->
     <!-- two config sections: first binder, second local
         switch which one has type="text/x-thebe-config"
