+++ examples/demo.html
@@ -1,14 +1,19 @@
 <html>
   <head>
     <title>Thebe demo</title>
+    <link rel="stylesheet" type="text/css" href="index.css" />
   </head>
   <body>
     <h1>Thebe demo</h1>
 
     <ul>
-      <li><a href="demo-static.html">A static web page</li>
+      <li><a href="demo-static.html">A static web page without Thebe</li>
       <li><a href="demo-active.html">The same, active</li>
       <li><a href="demo-active-variation.html">Variation</li>
+      <li><a href="demo-preview.html">Pre-computed cell output</li>
+      <li><a href="demo-activate-button.html">Built in activate button</li>
+      <li><a href="demo-status-widget.html">Built in status field</li>
+      <li><a href="demo-launch-button.html">Example of a custom launch &apos; load script</li>
     </ul>
 
     <h2>References</h2>
