+++ examples/index.html
@@ -69,32 +69,26 @@ plt.plot(x, np.sin(x))
 plt.plot(x, np.cos(x))</pre
     >
 
-    <h2>More examples</h2>
-
+    <h2>More examples &amp; demos</h2>
     <ul>
       <li>
-        <a href="widgets.html">Making use of Jupyter interactive widgets</a>;
+        <a href="local.html">Using a local Jupyter server as kernel provider</a>
+      </li>
+      <li>
+        <a href="widgets.html">Making use of Jupyter interactive widgets</a>
       </li>
       <li>
         <a href="prompts.html"
           >Alternative computational environments; code cells with prompts and
           outputs</a
-        >;
+        >
       </li>
       <li>
-        <a href="local.html">Using a local Jupyter server as kernel provider</a
-        >;
+        <a href="demo-preview.html">Setting predefined output for cells</a>
       </li>
+      <li><a href="demo-activate-button.html">Simple activate button</a></li>
       <li>
-        <a href="demo-preview.html">Setting predefined output for cells</a>;
-      </li>
-      <li><a href="demo-activate-button.html">Simple activate button</a>;</li>
-      <li>
-        <a href="demo-activate-button.html">Using the Status Widget on a page</a
-        >;
-      </li>
-      <li>
-        <a href="status/status_field.html">Simple status field and styling</a>;
+        <a href="demo-status-widget.html">Using the Status Widget on a page</a>
       </li>
       <li>
         <a
