--- examples/status/thebe_status_field.css
@@ -1,36 +1,0 @@
-.thebe-status-field {
-  background-color: #ccc;
-  display: inline-block;
-  margin: auto;
-  padding: 2px 8px;
-}
-
-.thebe-status-field.thebe-status-building {
-  /* Binder is building the image */
-  background-color: yellow;
-}
-
-.thebe-status-field.thebe-status-launching {
-  /* Binder launching the server */
-  background-color: yellow;
-}
-
-.thebe-status-field.thebe-status-starting {
-  /* Binder starting the server */
-  background-color: yellow;
-}
-
-.thebe-status-field.thebe-status-ready {
-  /* The kernel is connected and ready */
-  background-color: lightgreen;
-}
-
-.thebe-status-field.thebe-status-failed {
-  /* Building the image or launching the kernel failed */
-  background-color: red;
-}
-
-.thebe-status-field.thebe-status-busy {
-  /* The kernel is busy (this event is not hooked up yet) */
-  background-color: yellow;
-}
