+++ examples/widgets.html
@@ -1,42 +1,46 @@
 <html>
-<head>
-  <meta HTTP-EQUIV="Content-Type" CONTENT="text/html; charset=UTF-8">
-  <title>Thebe examples</title>
-  <link rel="stylesheet" type="text/css" href="index.css" />
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title>Thebe examples</title>
+    <link rel="stylesheet" type="text/css" href="index.css" />
 
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js">
-  </script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js"></script>
 
-  <!-- Configure and load Thebe !-->
-  <script type="text/x-thebe-config">
-    {
-      bootstrap: true,
-      binderOptions: {
-        repo: "jupyter-widgets/ipyleaflet",
-        ref: "0.9.0"
-      },
-      kernelOptions: {
-        name: "python3"
+    <!-- Configure and load Thebe !-->
+    <script type="text/x-thebe-config">
+      {
+        bootstrap: true,
+        binderOptions: {
+          repo: "jupyter-widgets/ipyleaflet",
+          ref: "0.9.0"<link rel="stylesheet" type="text/css" href="index.css" />
+        },
+        kernelOptions: {
+          name: "python3"
+        }
       }
-    }
-  </script>
-  <script type="text/javascript"
-      src="https://unpkg.com/thebe@latest"></script>
-  <!-- or to use a local build: -->
-  <!-- <script type="text/javascript" src="../lib/index.js"></script> -->
+    </script>
+    <script
+      type="text/javascript"
+      src="https://unpkg.com/thebe@latest"
+    ></script>
+    <!-- or to use a local build: -->
+    <!-- <script type="text/javascript" src="../lib/index.js"></script> -->
+  </head>
+  <body>
+    <h1>Thebe and Jupyter interactive widgets</h1>
 
-</head>
-<body>
-  <h1>Thebe and Jupyter interactive widgets</h1>
+    <p>
+      This page illustrates a minimal setup to get custom Jupyter interactive
+      widgets in Thebe.
+    </p>
 
-  <p>This page illustrates a minimal setup to get custom Jupyter interactive
-  widgets in Thebe.</p>
+    <p>
+      This requires adding a script tag to load requirejs on page, for thebelab
+      to be able to load the JavaScript assets for custom widgets.
+    </p>
 
-  <p>This requires adding a script tag to load requirejs on
-  page, for thebelab to be able to load the JavaScript assets for custom
-  widgets.</p>
-
-  <pre data-executable="true" data-language="python">from ipyleaflet import *
+    <pre data-executable="true" data-language="python">
+from ipyleaflet import *
 from ipywidgets import *
 
 # A leaflet map with the Strava tile layer.
@@ -54,7 +58,7 @@ def handle_interaction(**kwargs):
     if kwargs.get('type') == 'mousemove':
         coordinates.value = str(kwargs.get('coordinates'))
 
-m.on_interaction(handle_interaction)</pre>
-
-</body>
+m.on_interaction(handle_interaction)</pre
+    >
+  </body>
 </html>
