+++ src/activate.css
@@ -1,0 +1,6 @@
+.thebe-activate-button {
+  display: inline-block;
+  width: 120px;
+  height: 45px;
+  font-size: 1.5em;
+}
