+++ src/index.js
@@ -7,9 +7,14 @@ if (typeof window !== "undefined") {
   window.thebelab = thebelab;
   window.thebelab.$ = $;
   const options = thebelab.mergeOptions();
-  if (options["mountStatusWidget"]) {
+  if (options.mountStatusWidget) {
     document.addEventListener("DOMContentLoaded", () => {
       thebelab.mountStatusWidget();
+    });
+  }
+  if (options.mountActivateWidget) {
+    document.addEventListener("DOMContentLoaded", () => {
+      thebelab.mountActivateWidget();
     });
   }
   if (options["bootstrap"]) {
